Enh 7 html element (#21)
* Issue #7: HtmlElement POC

Adding HtmlElement with getter for attributes of element. as well as a usage example within GroupComponent Test.
As part of this a new Execute JS Action was created.

This is to serve as an example and a starting point for HtmlElement and implementing the properties.
Providing getAttributes functionality with unit tests and a test page.